user-manager cli v1 抽取grant子命令
